First functionality

+++ Antikythera/ATKIDisplay.cpp
@@ -48,25 +48,25 @@ ATKColor::RGBA* ATKIDisplay::framebuffer(uint8_t layer) {
 	return NULL;
 }
 
-void ATKIDisplay::circle(int8_t x, int8_t y, int8_t r, ATKColor::HSVA c, int8_t thickness, uint8_t style, uint8_t layer) {
+void ATKIDisplay::circle(int16_t x, int16_t y, int16_t r, ATKColor::HSVA c, int16_t thickness, uint8_t style, uint8_t layer) {
 }
 
 
-void ATKIDisplay::circle(int8_t x, int8_t y, int8_t r, ATKColor::RGBA c, int8_t thickness, uint8_t style, uint8_t layer) {
+void ATKIDisplay::circle(int16_t x, int16_t y, int16_t r, ATKColor::RGBA c, int16_t thickness, uint8_t style, uint8_t layer) {
 }
 
 
-void ATKIDisplay::line(int8_t x1, int8_t y1, int8_t x2, int8_t y2, ATKColor::HSVA c, int8_t thickness, uint8_t style, uint8_t mode, uint8_t layer) {
+void ATKIDisplay::line(int16_t x1, int16_t y1, int16_t x2, int16_t y2, ATKColor::HSVA c, int16_t thickness, uint8_t style, uint8_t mode, uint8_t layer) {
 }
 
 
-void ATKIDisplay::line(int8_t x1, int8_t y1, int8_t x2, int8_t y2, ATKColor::RGBA c, int8_t thickness, uint8_t style, uint8_t mode, uint8_t layer) {
+void ATKIDisplay::line(int16_t x1, int16_t y1, int16_t x2, int16_t y2, ATKColor::RGBA c, int16_t thickness, uint8_t style, uint8_t mode, uint8_t layer) {
 }
 
 
-void ATKIDisplay::point(int8_t x, int8_t y, ATKColor::HSVA c, uint8_t style, uint8_t layer) {
+void ATKIDisplay::point(int16_t x, int16_t y, ATKColor::HSVA c, uint8_t style, uint8_t layer) {
 }
 
 
-void ATKIDisplay::point(int8_t x, int8_t y, ATKColor::RGBA c, uint8_t style, uint8_t layer) {
+void ATKIDisplay::point(int16_t x, int16_t y, ATKColor::RGBA c, uint8_t style, uint8_t layer) {
 }
